third_party/blink/web_tests/accessibility/input-type-range-value-change-expected.txt

Check whether changing slider's value without step attribute works properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS range1.intValue is 50
PASS range1.intValue is 51
PASS range1.intValue is 52
PASS range1.intValue is 51
PASS range1.intValue is 52
PASS range2.intValue is 5
PASS range2.intValue is 6
PASS range2.intValue is 7
PASS range2.intValue is 6
PASS range2.intValue is 7
PASS range3.intValue is 10
PASS range3.intValue is 15
PASS range3.intValue is 20
PASS range3.intValue is 15
PASS range3.intValue is 20
PASS successfullyParsed is true

TEST COMPLETE

  
